// src/app/shared/interfaces/item.ts
export interface Item {
    itemId: string;
    itemName: string;
    manufacturer: string;
    description: string;
    materialInformation: string;
    gender: string;
    productType: string;
    sleeve: string;
    leg: string;
    collar: string;
    manufacture: string;
    bagType: string;
    gramWeight: string;
    material: string;
    countryOfOrigin: string;
    image: string;
}